backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.7.0
pydantic-settings==2.2.1
gspread==6.1.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
pandas==2.2.2
openpyxl==3.1.2
reportlab==4.1.0
Pillow==10.3.0
qrcode==7.4.2
APScheduler==3.10.4
httpx==0.27.0
python-dotenv==1.0.1
aiofiles==23.2.1
